An initial kludge-y hack of Value Error
This initial hack unfortunately moves the value error from an always-value error to a sometimes value-error in the ROC update function in line 303, in the BlackLittermanSymbolData class. What we gain is occasionally backtesting may complete successfully over certain dates. Also, this has not been tested since the removal of a try/ except block around line 303.